Change heatmap scale
Changed the heatmap scale such that it shows a value of "0" when there's
no emotion for that particular spot in the page and a value of "3" when
there's a strong emotion for a particular point in the page.

This is opposed to the previous values of "null" for no emotion and "2"
for strong emotion, that brought to the need of "interpolating" to plot
the missing values.