Fix page duplication error
Fixed an error due to which pages with an anchor were considered
different. Now, all of the following pages will be considered the same
page:

https://example.com:80/
https://example.com:80/#anchor

As they'll be stored as "https://example.com:80/". Note the lack of the
anchor. Note that example.com:80 and example.com are still saved as
different pages, and it's the same by varying query parameters or adding
or removing the final slash.

FIXME: example.com, example.com/, and example.com/index.html are saved
       as different pages, but they are always the same page.